fix column type start_at end_at for activity

+++ src/entity/Activity.ts
@@ -39,11 +39,11 @@ export class Activity {
   })
   description!: string;
   @Column({
-    type: "bigint",
+    type: "int",
   })
   start_at!: number;
   @Column({
-    type: "bigint",
+    type: "int",
   })
   end_at!: number;
   @Column({
